manually enable services
seems that this somehow still runs an nginx (maybe because of the srvos
mixin? TODO debug that) but it only listens locally and I really plan on
running nginx anyways so it's probably fine. this will make it a lot
easier to turn on/off stuff than manually commenting out include lines,
I think this is the way things are "supposed" to be done ya