wip: qemu networking
the install process is taking a bit because turns out running a VM
(qemu) inside a VM (WSL), while on battery, while the innermost VM is
swapping, is pretty slow lol. plus the filesystem performance of all
that isn't great anyways. ah well nothing do to but to wait for it to
finish ig