wip: qemu-aarch64 configuration
I have confirmed it's a problem with nixos-rebuild-ng trying to run
aarch64 binaries on my x86_64 system. The workaround is to get it to not
"reexec" (run with the nixpkgs of the target system? (but it doesn't
even seem to be running on the target system??)) with the `--fast` flag.
I would like to report this issue with a minimal testcase before
applying this solution to my situation.